+++ examples/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @naive-potato-ui/examples
 
+## 1.0.46
+
+### Patch Changes
+
+- add parse fn
+
+- Updated dependencies []:
+  - @naive-potato-ui/ui@1.0.45
+
 ## 1.0.45
 
 ### Patch Changes