+++ packages/custom-upload/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @naive-potato-ui/custom-upload
 
+## 1.0.11
+
+### Patch Changes
+
+- add parse fn
+
 ## 1.0.10
 
 ### Patch Changes